refactor: verify.ts and verifyMultipleAggregateSignatures.ts
