+++ blst-bun/src/signature.ts
@@ -1,7 +1,5 @@
-import { write } from "bun";
 import { binding, writeReference } from "./binding";
-import { BLST_SUCCESS, MAX_SIGNATURE_SETS_PER_JOB, SIGNATURE_LENGTH_COMPRESSED, SIGNATURE_LENGTH_UNCOMPRESSED } from "./const";
-import { writePublicKeysReference, type PublicKey } from "./publicKey";
+import { BLST_SUCCESS, SIGNATURE_LENGTH_COMPRESSED, SIGNATURE_LENGTH_UNCOMPRESSED } from "./const";
 import { blstErrorToReason, fromHex, toHex } from "./util";
 
 export class Signature {
@@ -93,129 +91,4 @@ export class Signature {
     writeReference(this.blst_point, out, offset);
   }
 
-}
-
-export interface SignatureSet {
-  msg: Uint8Array;
-  pk: PublicKey;
-  sig: Signature;
-};
-
-// global pairing buffer to be reused across multiple calls
-const pairing = new Uint8Array(binding.sizeOfPairing());
-
-/**
- * Verify a signature against a message and public key.
- *
- * If `pk_validate` is `true`, the public key will be infinity and group checked.
- *
- * If `sig_groupcheck` is `true`, the signature will be group checked.
- */
-export function verify(msg: Uint8Array, pk: PublicKey, sig: Signature, pkValidate?: boolean | undefined | null, sigGroupcheck?: boolean | undefined | null): boolean {
-  const res = binding.verifySignature(sig.blst_point, sigGroupcheck ?? false, msg, msg.length, pk.blst_point, pkValidate ?? false);
-  return res === 0;
-}
-
-/**
- * Verify an aggregated signature against multiple messages and multiple public keys.
- *
- * If `pk_validate` is `true`, the public keys will be infinity and group checked.
- *
- * If `sigs_groupcheck` is `true`, the signatures will be group checked.
- *
- * The down side of zig binding is all messages have to be the same length.
- */
-export function aggregateVerify(msgs: Array<Uint8Array>, pks: Array<PublicKey>, sig: Signature, pkValidate?: boolean | undefined | null, sigsGroupcheck?: boolean | undefined | null): boolean {
-  if (msgs.length < 1) {
-    throw new Error("At least one message is required");
-  }
-  if (msgs.length !== pks.length) {
-    throw new Error("Number of messages must be equal to the number of public keys");
-  }
-
-  const msgLen = msgs[0].length;
-  for (let i = 1; i < msgs.length; i++) {
-    if (msgs[i].length !== msgLen) {
-      throw new Error("All messages must be the same length");
-    }
-  }
-
-  const msgs_ref = writeMessagesReference(msgs);
-  const pks_references = writePublicKeysReference(pks);
-  const res = binding.aggregateVerify(sig.blst_point, sigsGroupcheck ?? false, msgs_ref, msgs.length, msgLen, pks_references, pks.length, pkValidate ?? false, pairing, pairing.length);
-  return res == 0;
-}
-
-/**
- * Verify an aggregated signature against a single message and multiple public keys.
- *
- * Proof-of-possession is required for public keys.
- *
- * If `sigs_groupcheck` is `true`, the signatures will be group checked.
- */
-export function fastAggregateVerify(msg: Uint8Array, pks: Array<PublicKey>, sig: Signature, sigsGroupcheck?: boolean | undefined | null): boolean {
-  const pks_references = writePublicKeysReference(pks);
-  const res = binding.fastAggregateVerify(sig.blst_point, sigsGroupcheck ?? false, msg, msg.length, pks_references, pks.length, pairing, pairing.length);
-  return res === 0;
-}
-
-// global signature sets reference to be reused across multiple calls
-// each 2 tems are 8 bytes, store the reference of each signature set
-const signature_sets_ref = new Uint32Array(MAX_SIGNATURE_SETS_PER_JOB * 2);
-
-/**
- * Verify multiple aggregated signatures against multiple messages and multiple public keys.
- *
- * If `pks_validate` is `true`, the public keys will be infinity and group checked.
- *
- * If `sigs_groupcheck` is `true`, the signatures will be group checked.
- *
- * See https://ethresear.ch/t/fast-verification-of-multiple-bls-signatures/5407
- */
-export function verifyMultipleAggregateSignatures(sets: SignatureSet[], pksValidate?: boolean | undefined | null, sigsGroupcheck?: boolean | undefined | null): boolean {
-  if (sets.length > MAX_SIGNATURE_SETS_PER_JOB) {
-    throw new Error(`Number of signature sets exceeds the maximum of ${MAX_SIGNATURE_SETS_PER_JOB}`);
-  }
-
-  writeSignatureSetsReference(sets, signature_sets_ref.subarray(0, sets.length * 2));
-  const msgLength = 32;
-  const res = binding.verifyMultipleAggregateSignatures(signature_sets_ref, sets.length, msgLength, pksValidate ?? false, sigsGroupcheck ?? false, pairing, pairing.length);
-  return res === 0;
-}
-
-// global signature set data to be reused across multiple calls
-// each 6 items are 24 bytes, store 3 references of each signature set (msg + pk + sig)
-const signature_sets_data = new Uint32Array(MAX_SIGNATURE_SETS_PER_JOB * 6);
-
-function writeSignatureSetsReference(sets: SignatureSet[], out: Uint32Array): void {
-  let offset = 0;
-  for (const [i, set] of sets.entries()) {
-    writeSignatureSetReference(set, signature_sets_data, offset + i * 6);
-    // write pointer
-    writeReference(signature_sets_data.subarray(i * 6, i * 6 + 6), out, i * 2);
-  }
-}
-
-// each SignatureSet needs 24 bytes = 6 * uint32 for references
-function writeSignatureSetReference(set: SignatureSet, out: Uint32Array, offset: number): void {
-  writeReference(set.msg, out, offset);
-  set.pk.writeReference(out, offset + 2);
-  set.sig.writeReference(out, offset + 4);
-}
-
-
-const MAX_MSGS = 128;
-// global messages references to be reused across multiple calls
-const messages_refs = new Uint32Array(MAX_MSGS * 2);
-
-function writeMessagesReference(msgs: Uint8Array[]): Uint32Array {
-  if (msgs.length > MAX_MSGS) {
-    throw new Error(`Too many messages, max is ${MAX_MSGS}`);
-  }
-
-  for (let i = 0; i < msgs.length; i++) {
-    writeReference(msgs[i], messages_refs, i * 2);
-  }
-
-  return messages_refs.subarray(0, msgs.length * 2);
 }
